Add ubuntu test for java 11
