+++ molecule/resources/tests/test_jdk.py
@@ -1,0 +1,20 @@
+import os
+
+import testinfra.utils.ansible_runner
+
+testinfra_hosts = testinfra.utils.ansible_runner.AnsibleRunner(
+    os.environ['MOLECULE_INVENTORY_FILE']).get_hosts('java-jdk')
+
+
+def test_jre_version(host):
+    assert host.exists('java')
+    c = host.run('java -version')
+    assert c.rc == 0
+    assert c.stderr.startswith('openjdk version "1.8.0')
+
+
+def test_jdk_version(host):
+    assert host.exists('javac')
+    c = host.run('javac -version')
+    assert c.rc == 0
+    assert c.stderr.startswith('javac 1.8.0')
